chore: separate category view and edit page
