+++ assets/l10n/en_US.json
@@ -9,6 +9,7 @@
     "general.delete.permanentWarning": "This action is irreversible",
     "general.delete.confirmName": "Confirm deleting {name}?",
     "general.areYouSure": "Are you sure?",
+    "general.edit": "Edit",
     "general.save": "Save",
     "general.done": "Done",
     "general.confirm": "Confirm",
