+++ assets/l10n/mn_MN.json
@@ -9,6 +9,7 @@
     "general.delete.permanentWarning": "Энэ үйлдлийг буцаах боломжгүй",
     "general.delete.confirmName": "{name}-г устгах уу?",
     "general.areYouSure": "Итгэлтэй байна уу?",
+    "general.edit": "Засварлах",
     "general.save": "Хадгалах",
     "general.done": "Болсон",
     "general.confirm": "Батлах",
@@ -101,7 +102,7 @@
     "category": "Ангилал",
     "category.name": "Нэр",
     "category.new": "Ангилал үүсгэх",
-    "category.delete": "Ангилалыг устгах",
+    "category.delete": "Ангиллыг устгах",
     "category.delete.warning": "Энэ ангиллыг устгавал холбоотой {transactionCount} гүйлгээг ангилалгүй болохыг анхаарна уу. Энэ үйлдлийг буцаах боломжгүй юм!",
     "category.skip": "Ангилалгүй",
     "category.none": "Ангилалгүй",
